+++ vacefron/__init__.py
@@ -1,3 +1,3 @@
 from .client import Client
-from .errors import BadRequest, InternalServerError, NotFound, HTTPException
+from .errors import BadRequest, HTTPException, InternalServerError, NotFound
 from .image import Image