+++ changelog.md
@@ -1,5 +1,9 @@
 # VACEfron.py | Changelog
 See here what changed or broke each version.
 
+## v1.0.1
+- Added an optional `session` param to `vacefron.Client()`.
+- Renamed `_session` to `session`
+
 ## v1.0.0
 First release, yes.